Extract shinjo archer cost
